factored out small sets of operators from math
