+++ lang/python/python.py
@@ -136,7 +136,6 @@ ctx.lists["user.python_exception"] = {
 
 @ctx.action_class("user")
 class UserActions:
-    def code_operator_lambda():                actions.auto_insert('')
     def code_operator_subscript():
         actions.insert('[]')
         actions.key('left')
