+++ lang/python/python.talon
@@ -4,12 +4,15 @@ mode: command
 and mode: user.auto_lang
 and code.language: python
 -
-tag(): user.code_imperative
-tag(): user.code_operators
+tag(): user.code_base_imperative
+tag(): user.code_operators_array
 tag(): user.code_operators_assignment
+tag(): user.code_operators_bitwise
+tag(): user.code_operators_math
 tag(): user.code_comment
 tag(): user.code_gui_functions
 tag(): user.code_gui_libraries
+
 settings():
     user.code_private_function_formatter = "SNAKE_CASE"
     user.code_protected_function_formatter = "SNAKE_CASE"
