+++ lang/ruby/ruby.talon
@@ -4,10 +4,14 @@ mode: command
 and mode: user.auto_lang
 and code.language: ruby
 -
-tag(): user.code_imperative
-tag(): user.code_operators
+tag(): user.code_base_imperative
+tag(): user.code_operators_array
 tag(): user.code_operators_assignment
+tag(): user.code_operators_bitwise
+tag(): user.code_operators_lambda
+tag(): user.code_operators_math
 tag(): user.code_comment
+
 settings():
     user.code_private_function_formatter = "SNAKE_CASE"
     user.code_protected_function_formatter = "SNAKE_CASE"
@@ -22,6 +26,7 @@ args pipe:
     insert("||")
     key(left)
 
+# TODO: duplicate from tag code_base_imperative?
 state do: "do "
 state end: "end"
 state begin: "begin"
