+++ lang/tags/operators/array.talon
@@ -1,0 +1,5 @@
+tag: user.code_operators_array
+-
+
+# array subscription
+op subscript: user.code_operator_subscript()
