+++ lang/tags/operators/bitwise.talon
@@ -1,0 +1,12 @@
+tag: user.code_operators_bitwise
+-
+
+#bitwise operators
+[op] bitwise and: user.code_operator_bitwise_and()
+[op] bitwise or: user.code_operator_bitwise_or()
+
+# TODO: split these out into separate logical and bitwise operator commands
+
+(op | logical | bitwise) (ex | exclusive) or: user.code_operator_bitwise_exclusive_or()
+(op | logical | bitwise) (left shift | shift left): user.code_operator_bitwise_left_shift()
+(op | logical | bitwise) (right shift | shift right): user.code_operator_bitwise_right_shift()
