+++ lang/tags/operators/collection.talon
@@ -1,0 +1,6 @@
+tag: user.code_operators_collection
+-
+
+# NOTE: Only used in the R language. Perhaps this doesn't need its own tag?
+
+(op | is) in: user.code_operator_in()
