+++ lang/tags/operators/lambda.py
@@ -1,0 +1,14 @@
+from talon import Context, Module
+
+ctx = Context()
+mod = Module()
+
+# TODO: this probably shouldn't be in operators
+
+mod.tag("code_operators_lambda", desc="Tag for enabling commands for anonymous functions")
+
+@mod.action_class
+class Actions:
+
+    def code_operator_lambda():
+        """code_operator_lambda"""
