+++ lang/tags/operators/math.py
@@ -3,18 +3,13 @@ from talon import Context, Module
 ctx = Context()
 mod = Module()
 
-# TODO: rename this tag to 'code_operator' for consistency?
-mod.tag("code_operators", desc="Tag for enabling generic operator commands")
 
+# TODO: Could split into numeric, comparison, and logic?
+
+mod.tag("code_operators_math", desc="Tag for enabling mathematical operator commands")
 
 @mod.action_class
 class Actions:
-
-    def code_operator_lambda():
-        """code_operator_lambda"""
-
-    def code_operator_subscript():
-        """code_operator_subscript (e.g., C++ [])"""
 
     def code_operator_subtraction():
         """code_operator_subtraction"""
@@ -52,26 +47,8 @@ class Actions:
     def code_operator_less_than_or_equal_to():
         """code_operator_less_than_or_equal_to"""
 
-    def code_operator_in():
-        """code_operator_in"""
-
     def code_operator_and():
         """code_operator_and"""
 
     def code_operator_or():
         """code_operator_or"""
-
-    def code_operator_bitwise_and():
-        """code_operator_bitwise_and"""
-
-    def code_operator_bitwise_or():
-        """code_operator_bitwise_or"""
-
-    def code_operator_bitwise_exclusive_or():
-        """code_operator_bitwise_exclusive_or"""
-
-    def code_operator_bitwise_left_shift():
-        """code_operator_bitwise_left_shift"""
-
-    def code_operator_bitwise_right_shift():
-        """code_operator_bitwise_right_shift"""
