+++ lang/tags/operators/pointer.py
