+++ lang/tags/operators/pointer.talon
