+++ lang/typescript/typescript.talon
@@ -4,9 +4,12 @@ mode: command
 and mode: user.auto_lang
 and code.language: typescript
 -
-tag(): user.code_imperative
-tag(): user.code_operators
+tag(): user.code_base_imperative
+tag(): user.code_operators_array
 tag(): user.code_operators_assignment
+tag(): user.code_operators_bitwise
+tag(): user.code_operators_lambda
+tag(): user.code_operators_math
 tag(): user.code_comment
 tag(): user.code_gui_functions
