+++ lang/vimscript/vimscript.talon
@@ -4,10 +4,11 @@ mode: command
 and mode: user.auto_lang
 and code.language: vimscript
 -
-tag(): user.code_imperative
-tag(): user.code_operators
+tag(): user.code_base_imperative
 tag(): user.code_operators_assignment
+tag(): user.code_operators_math
 tag(): user.code_comment
+
 # XXX - revisit these
 settings():
     user.code_private_function_formatter = "SNAKE_CASE"
